Create EventListener class
